Update list.R
Revised lines 77-82, replacing single bracket notation to double bracket notation for subsetting columns to accomdate both `tibble` and `data.frame` objects. The evaluation `is.character(edges[, 1])` for a tibble evaluates the `tibble` object and not the column. I revised this to `is.character(edges[[1]])`, which evaluates the column. This works for `data.frame` and `tibble` objects. Each case of single bracket notation was replaced with double bracket notation.